Update LispBM.lsp
optOffDelay is the solution for smooth cruisecontrol powering the motor.  Unfortunately there isn't that option for set-duty.  I haven't gotten set-current working in testing with no load.  Maybe it works with load.  set-current-rel works.  But this code uses get-duty to get a relevant 0.0 - 1.0 value to set the relative current to.  Which isn't 1 to 1.  This works but needs work.